[CSB] Adicionar e centraliza alguns comportamentos em comuns nos testes; Modificação dos tipos do objeto pageable long -> int

+++ app/src/main/java/br/com/casasbahia/converter/impl/SellerConverterImpl.java
@@ -58,7 +58,7 @@ public class SellerConverterImpl
             sellerPage.getSize(),
             sellerPage.getNumberOfElements(),
             sellerPage.getTotalPages(),
-            sellerPage.getTotalElements() );
+            (int) sellerPage.getTotalElements() );
         return new SellerPageableDTO( pageableDTO, toSellerDTO( sellerPage.getContent() ) );
     }
